# swagger-spec.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/services/{id}/status": {
      "put": {
        "operationId": "updateStatus",
        "summary": "Update service status",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Service status updated successfully"
          }
        },
        "tags": [
          "Services"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/services/{id}/featured": {
      "put": {
        "operationId": "toggleFeatured",
        "summary": "Toggle service featured status",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Service featured status updated"
          }
        },
        "tags": [
          "Services"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/categories": {
      "post": {
        "operationId": "create",
        "summary": "Create new category",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCategoryDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Category created successfully"
          }
        },
        "tags": [
          "Categories"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "get": {
        "operationId": "findAll",
        "summary": "Get all categories",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Return all categories"
          }
        },
        "tags": [
          "Categories"
        ]
      }
    },
...
    "/categories/{id}": {
      "get": {
        "operationId": "findOne",
        "summary": "Get category by id",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Return category details"
          }
        },
        "tags": [
          "Categories"
        ]
      },
      "put": {
        "operationId": "update",
        "summary": "Update category",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCategoryDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Category updated successfully"
          }
        },
        "tags": [
          "Categories"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "delete": {
        "operationId": "remove",
        "summary": "Delete category",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Category deleted successfully"
          }
        },
        "tags": [
          "Categories"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/categories/{id}/featured": {
      "put": {
        "operationId": "toggleFeatured",
        "summary": "Toggle category featured status",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Category featured status updated"
          }
        },
        "tags": [
          "Categories"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
    "/payments/{id}/status": {
      "put": {
        "operationId": "updateStatus",
        "summary": "Update payment status (Admin)",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdatePaymentStatusDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Payment status updated"
          }
        },
        "tags": [
          "Payments"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
    "/reviews/{id}/status": {
      "put": {
        "operationId": "updateStatus",
        "summary": "Update review status (Admin)",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateReviewStatusDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Review status updated successfully"
          }
        },
        "tags": [
          "Reviews"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
    "/notifications": {
      "get": {
        "operationId": "findAll",
        "summary": "Get user notifications",
        "parameters": [
          {
            "name": "status",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Return user notifications"
          }
        },
        "tags": [
          "Notifications"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "delete": {
        "operationId": "removeAll",
        "summary": "Delete all notifications of a user (Admin)",
        "parameters": [
          {
            "name": "userId",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "All notifications deleted successfully"
          }
        },
        "tags": [
          "Notifications"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/notifications/unread-count": {
      "get": {
        "operationId": "getUnreadCount",
        "summary": "Get unread notifications count",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Return unread notifications count"
          }
        },
        "tags": [
          "Notifications"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/notifications/{id}": {
      "get": {
        "operationId": "findOne",
        "summary": "Get notification details",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Return notification details"
          }
        },
        "tags": [
          "Notifications"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "delete": {
        "operationId": "remove",
        "summary": "Delete notification (Admin)",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Notification deleted successfully"
          }
        },
        "tags": [
          "Notifications"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/notifications/{id}/status": {
      "put": {
        "operationId": "updateStatus",
        "summary": "Update notification status",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateNotificationStatusDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Notification status updated"
          }
        },
        "tags": [
          "Notifications"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/notifications/mark-all-read": {
      "put": {
        "operationId": "markAllAsRead",
        "summary": "Mark all notifications as read",
        "parameters": [],
        "responses": {
          "200": {
            "description": "All notifications marked as read"
          }
        },
        "tags": [
          "Notifications"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
  },
  "info": {
    "title": "HizmetBul API",
    "description": "HizmetBul Servis Platformu API Dokümantasyonu",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "CreateCategoryDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Ev Temizliği"
          },
          "slug": {
            "type": "string",
            "example": "ev-temizligi"
          },
          "description": {
            "type": "string",
            "example": "Profesyonel ev temizlik hizmetleri"
          },
          "icon": {
            "type": "string",
            "example": "fas fa-home"
          },
          "image": {
            "type": "string",
            "example": "temizlik.jpg"
          },
          "parentId": {
            "type": "string",
            "example": "507f1f77bcf86cd799439011"
          },
          "order": {
            "type": "number",
            "example": 1
          },
          "isActive": {
            "type": "boolean",
            "example": true
          },
          "isFeatured": {
            "type": "boolean",
            "example": false
          },
          "metadata": {
            "type": "object"
          }
        },
        "required": [
          "name",
          "slug"
        ]
      },
      "UpdateCategoryDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Ev Temizliği"
          },
          "slug": {
            "type": "string",
            "example": "ev-temizligi"
          },
          "description": {
            "type": "string",
            "example": "Profesyonel ev temizlik hizmetleri"
          },
          "icon": {
            "type": "string",
            "example": "fas fa-home"
          },
          "image": {
            "type": "string",
            "example": "temizlik.jpg"
          },
          "parentId": {
            "type": "string",
            "example": "507f1f77bcf86cd799439011"
          },
          "order": {
            "type": "number",
            "example": 1
          },
          "isActive": {
            "type": "boolean",
            "example": true
          },
          "isFeatured": {
            "type": "boolean",
            "example": false
          },
          "metadata": {
            "type": "object"
          }
        }
      },
...
      "UpdatePaymentStatusDto": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "pending",
              "processing",
              "completed",
              "failed",
              "refunded"
            ]
          },
          "transactionId": {
            "type": "string",
            "example": "TR123456789"
          },
          "error": {
            "type": "string",
            "example": "Payment failed: Insufficient funds"
          },
          "notes": {
            "type": "string"
          }
        },
        "required": [
          "status"
        ]
      },
...
      "UpdateReviewStatusDto": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "pending",
              "approved",
              "rejected"
            ]
          },
          "reason": {
            "type": "string"
          }
        },
        "required": [
          "status"
        ]
      },
...
      "UpdateNotificationStatusDto": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "unread",
              "read",
              "archived"
            ]
          }
        },
        "required": [
          "status"
        ]
      },
...
    }
  }
}